pythonWorker/requirements.txt
setuptools
pyzmq
pymilvus
onnxruntime_genai